backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
openai
chromadb
python-docx
pypdf2
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
spacy
lxml
spacy[lookups]
python-dotenv
langchain_openai
langchain-community
langchain-chroma